MCMRSimulator.jl API

This is the API for MCMRSimulator. For a more friendly introduction, click here

MCMRSimulator.MCMRSimulator

MCMRSimulator.Siemens_Connectom

MCMRSimulator.Siemens_Prisma

MCMRSimulator.Siemens_Terra

MCMRSimulator.gyromagnetic_ratio

MCMRSimulator.Annulus

MCMRSimulator.BaseObstruction

MCMRSimulator.BoundingBox

MCMRSimulator.Collision

MCMRSimulator.Cylinder

MCMRSimulator.Field

MCMRSimulator.FixedXoshiro

MCMRSimulator.Float

MCMRSimulator.GridShape

MCMRSimulator.InstantGradient

MCMRSimulator.LinearGradients

MCMRSimulator.LocalEnvironment

MCMRSimulator.MRGradients

MCMRSimulator.Mesh

MCMRSimulator.Microstructure

MCMRSimulator.Movement

MCMRSimulator.Obstruction

MCMRSimulator.ObstructionProperties

MCMRSimulator.PlotPlane

MCMRSimulator.PosVector

MCMRSimulator.RFPulse

MCMRSimulator.Readout

MCMRSimulator.Scanner

MCMRSimulator.Sequence

MCMRSimulator.SequenceComponent

MCMRSimulator.Simulation

MCMRSimulator.Snapshot

MCMRSimulator.Sphere

MCMRSimulator.Spin

MCMRSimulator.SpinOrientation

MCMRSimulator.Spiral

MCMRSimulator.StepWiseGradients

MCMRSimulator.TransformObstruction

MCMRSimulator.Wall

Base.Libc.time

MCMRSimulator.annuli

MCMRSimulator.apply

MCMRSimulator.color

MCMRSimulator.corners

MCMRSimulator.correct_collisions

MCMRSimulator.create_gradients

MCMRSimulator.cylinders

MCMRSimulator.derive_qval_time

MCMRSimulator.detect_collision

MCMRSimulator.draw_step

MCMRSimulator.dwi

MCMRSimulator.dyad_snapshot

MCMRSimulator.dyad_snapshot!

MCMRSimulator.evolve

MCMRSimulator.evolve_to_time

MCMRSimulator.evolve_to_time

MCMRSimulator.expand

MCMRSimulator.field

MCMRSimulator.fit_time

MCMRSimulator.flip_angle

MCMRSimulator.get_gradient

MCMRSimulator.get_sequence

MCMRSimulator.get_times

MCMRSimulator.image_snapshot

MCMRSimulator.image_snapshot!

MCMRSimulator.isinside

MCMRSimulator.load_mesh

MCMRSimulator.longitudinal

MCMRSimulator.lorentz_off_resonance

MCMRSimulator.mesh_grid_intersection

MCMRSimulator.next_pulse

MCMRSimulator.normal

MCMRSimulator.off_resonance

MCMRSimulator.off_resonance

MCMRSimulator.orientation

MCMRSimulator.perfect_dwi

MCMRSimulator.phase

MCMRSimulator.phase

MCMRSimulator.plot_geometry

MCMRSimulator.plot_geometry!

MCMRSimulator.plot_off_resonance

MCMRSimulator.plot_off_resonance!

MCMRSimulator.plot_snapshot

MCMRSimulator.plot_trajectory2d

MCMRSimulator.plot_trajectory2d!

MCMRSimulator.plot_trajectory3d

MCMRSimulator.plot_trajectory3d!

MCMRSimulator.ply_from_mesh

MCMRSimulator.position

MCMRSimulator.position

MCMRSimulator.produces_off_resonance

MCMRSimulator.project

MCMRSimulator.project

MCMRSimulator.project_geometry

MCMRSimulator.project_on_grid

MCMRSimulator.random_annuli

MCMRSimulator.random_cylinders

MCMRSimulator.random_gauss

MCMRSimulator.random_on_sphere

MCMRSimulator.random_positions_radii

MCMRSimulator.random_spheres

MCMRSimulator.random_spirals

MCMRSimulator.ray_grid_intersections

MCMRSimulator.readout

MCMRSimulator.relax

MCMRSimulator.rotate_bvec

MCMRSimulator.scatter_snapshot

MCMRSimulator.scatter_snapshot!

MCMRSimulator.sequence_plot

MCMRSimulator.signal

MCMRSimulator.spheres

MCMRSimulator.spiral_theta

MCMRSimulator.spirals

MCMRSimulator.total_susceptibility

MCMRSimulator.trajectory

MCMRSimulator.transfer

MCMRSimulator.transverse

MCMRSimulator.walls

MakieCore.plot!

MCMRSimulator.MCMRSimulator — Module

This package supports the running of MR Monte Carlo simulations.

In these simulations hundreds of thousands or millions of particles randomly diffuse through some tissue microstructure constrained by a series of Obstruction objects. The spins of these particles will be evolved by the effect of one or more Sequence objects as well as spatially varying Field objecs describing the R1, R2, diffusivity, and off-resonance fields. Off-resonance fields can also be generated by the Obstruction objects themselves. All of these variables are combined into a single Simulation object. See Simulation for how to run the simulation.

Plotting support for the output is also available based on Makie.jl.

MCMRSimulator.Siemens_Connectom — Constant

Siemens 3T Connectom MRI scanner (Qiuyun Fan, Cornelius Eichner, Maryam Afzali, Lars Mueller, Chantal M. W. Tax, Mathias Davids, Mirsad Mahmutovic, Boris Keil, Berkin Bilgic, Kawin Setsompop, Hong-Hsi Lee, Qiyuan Tian, Chiara Maffei, Gabriel {Ramos-Llord{\'e}n}, Aapo Nummenmaa, Thomas Witzel, Anastasia Yendiki, Yi-Qiao Song, Chu-Chung Huang, Ching-Po Lin, Nikolaus Weiskopf, Alfred Anwander, Derek K. Jones, Bruce R. Rosen, Lawrence L. Wald, Susie Y. Huang (2022)).

MCMRSimulator.Siemens_Prisma — Constant

Siemens MAGNETOM 3T Prisma MRI scanner (https://www.siemens-healthineers.com/en-uk/magnetic-resonance-imaging/3t-mri-scanner/magnetom-prisma).

MCMRSimulator.Siemens_Terra — Constant

Siemens MAGNETOM 7T Terra MRI scanner (https://www.siemens-healthineers.com/en-uk/magnetic-resonance-imaging/7t-mri-scanner/magnetom-terra)

MCMRSimulator.gyromagnetic_ratio — Constant

gyromagnetic ratio in 10^6 rad/ms/T.

MCMRSimulator.Annulus — Type

Annulus(inner_radius, outer_radius; chi_I=-0.1, chi_A=-0.1, myelin=false)

Create an annulus with an inner and outer radius. Using annuli is recommended. Water can freely diffuse within the inner radius, and between the inner and outer radius, but not in between. Myelin can be added by setting myelin to true. Restrictions are only present at the inner and outer cylinder, not in the myelin in between.

MCMRSimulator.BaseObstruction — Type

Base obstruction type hindering free diffusion.

These obstructions are always perfectly aligned with the main axes and centered at the origin (except for Mesh). They can be moved/rotated/repeated by applying TransformObstruction.

The dimensionality N indicates the dimensionality of the input data (1 for Wall, 2 for Cylinder, 3 for Sphere or Mesh).

MCMRSimulator.BoundingBox — Type

BoundingBox(lower::PosVector, upper::PosVector)
BoundingBox(obstruction::Obstruction)
BoundingBox(obstructions)

Creates a bounding box containing one or more Obstruction. For infinitely repeated objects (using TransformObstruction) the bounding box of the central object is returned.

Check whether particles are inside using isinside.

MCMRSimulator.Collision — Type

Collision(distance, normal, properties; index=0, inside=false)

A detected collision along the movement.

Parameters

distance: number between 0 and 1 indicating the distance of the collision from the origin (0) to the destination point (1)

normal: normal of the obstruction at the collision site. To get correct reflection the normal should point in the direction of the incoming particle.

properties: ObstructionProperties of the obstruction the spin collided with.

index: Index of which triangle in Mesh got hit

inside: Whether the obstruction was hit on the inside or the outside. For a mesh triangle the outside is considered in the direction of the normal. For a wall the outside is in the positive direction.

MCMRSimulator.Cylinder — Type

Cylinder(radius; chi_I=-0.1, chi_A=-0.1, g_ratio=1)

Creates a hollow cylinder with a radius of radius micrometer (default 1 micrometer) at the given location (default: origin). Generate cylinders using cylinders. See Myelinated cylinders for an explanation of the myelin sheath.

MCMRSimulator.Field — Type

Describes the spatial distribution of the R1, R2, diffusivity, and off-resonance fields.

Construct using field.

MCMRSimulator.FixedXoshiro — Type

Immutable version of the Xoshiro random number generator state

Used to store the current state in the Spin object. To evolve the spin in a predictable manner set the seed using copy!(spin.rng, Random.TaskLocalRNG).

MCMRSimulator.Float — Type

Float type used during the simulation (default: Float64). This has to be changed in the code.

MCMRSimulator.GridShape — Type

GridShape(bounding_box::BoundingBox, size::SVector{3, Int})
GridShape(bounding_box::BoundingBox, size::Int)

Defines a 3D grid within the BoundingBox. size sets the number of voxels along each dimension. Setting size to a single integer will set that number along each dimension (except dimensions of infinite size, which are always just one voxel wide).

MCMRSimulator.InstantGradient — Type

InstantGradient(; qvec=[0, 0, 0], q_origin=0, time=0.)

Infinitely short gradient pulse that encodes phase information given by qvec and q_origin.

The phase offset at every position is given by qvec ⋅ position + q_origin.

The pulse is applied at given time (in milliseconds). every TR.

MCMRSimulator.LinearGradients — Type

create_gradients([(<start_time>, <gradient>), ...], interpolate=:linear)

Models MRI gradients as a sequence of linear functions.

MCMRSimulator.LocalEnvironment — Type

The off-resonance, R2, and R1 values at a single point in space

MCMRSimulator.MRGradients — Type

create_gradients([(<start_time>, <gradient>), ...], interpolate=:step/:linear)

Models MRI gradients as a sequence of stepwise or linear functions. Times are in ms and gradients in mT/m.

Off-resonance fields are estimated by calling get_gradient.

MCMRSimulator.Mesh — Type

Mesh(vertices, triangles)

An Obstruction formed from a triangular mesh.

MCMRSimulator.Microstructure — Type

Microstructure(; off_resonance=0., R2=0., R1=0., diffusivity=0., geometry=Obstruction[])

Describes the microstructure of the tissue.

This describes both the spatial distribution of the R1, R2, and off-resonance parameters (as Field objects), the diffusivity as a float, as well as the spatial geometry constraining the diffusion (as a sequence of Obstruction objects).

The R1, R2, and off-resonance parameters can be defined as one of:

value::Number: constant value across the microstructure.

(gradient::PosVector, value::Number): gradient across the microstructure.

field::Field: as generated using field.

MCMRSimulator.Movement — Type

Intermediate object used internally to represent a movement from one position to another

MCMRSimulator.Obstruction — Type

Supertype of any obstruction to the free diffusion of water. Some might also generate off-resonance fields.

There are two types of obstructions:

BaseObstruction with the basic obstructions (walls, sphere, cylinders, meshes)

TransformObstruction, which transform the base obstructions

MCMRSimulator.ObstructionProperties — Type

ObstructionProperties(kwargs...)
ObstructionProperties(obstruction)

The generic properties of an obstruction that all obstructions have in common. They include:

unique_identifier: unique identifier

MT_fraction: fraction of signal transfered in case of a collision

MCMRSimulator.PlotPlane — Type

Defines a finite plane in the 3D space used for plotting.

Constructor

PlotPlane(normal::PosVector=[0, 0, 1], position::PosVector=[0, 0, 0]; size=10., sizex=<size>, sizey=<size>, ngrid=100)

Arguments:

normal: length-3 vector with the orientation perpendicular to the plane (default to z-direction).

position: position of plane as a length-3 vector (defaults to origin).

sizex: size of the plane in the x-direction (before rotating to normal).

sizey: size of the plane in the y-direction (before rotating to normal).

size: set default value for sizex and sizey.

ngrid: number of grid elements to split the plane up into for plotting.

Spin projection onto plane

See project for details on how spins are projected onto the PlotPlane.

MCMRSimulator.PosVector — Type

General definition used for length-3 vectors. Mostly used for positions.

MCMRSimulator.RFPulse — Type

RFPulse(;time=0., flip_angle=0., phase=0.)

Instantaneous radio-frequency pulse that flips the spins by flip_angle degrees in a plane perpendicular to an axis in the x-y plane with an angle of phase degrees with respect to the x-axis at given time. Angles are in degrees (stored internally as radians) and the time is in milliseconds.

MCMRSimulator.Readout — Type

Readout(;time=0)

Readout the spins at the given time (in milliseconds) each TR

MCMRSimulator.Scanner — Type

Scanner(;B0=3., gradient=Inf, slew_rate=Inf)

Properties of an MRI scanner relevant for the MR signal simulations.

B0: magnetic field strength (in Tesla)

gradient: maximum gradient strength (in mT/m) long each axis

slew_rate: maximum rate of change in the gradient strength (in T/m/s)

MCMRSimulator.Sequence — Type

Sequence(;TR, gradients=nothing, pulses=nothing, B0=3., interplate_gradients=:step)

An MR sequence represented by a series of pulses repeated with a given repetition time (TR).

Possible sequence components are:

RFPulse: instantaneous radio-frequency pulse flipping the spin orientations.

InstantGradient: instantaneous gradients encoding spatial patterns in the spin phase distribution.

Readout: Store the spins at this timepoint.

The index of the next pulse is given by next_pulse(sequence, current_time). The time of this pulse can then be extracted using time(sequence, index). Note that these indices go on till infinite reflecting the repetitive nature of RF pulses over the TR time.

MCMRSimulator.SequenceComponent — Type

A single component of a longer MR Sequence.

MCMRSimulator.Simulation — Type

Simulation(sequences, [microstructure]; R1=0., R2=0., off_resonance=0., diffusivity=0., geometry=Obstruction[], store_every=5., timestep=0.5)

Defines the setup of the simulation and stores the output of the run.

Note the use of units:

all positions are in micrometers

all times are in milliseconds

magnetic fields are in Tesla

Arguments

sequences::AbstractVector{Sequence}: MR sequences to simulate. See Sequence.

microstructure::Microstructure: Description the tissue microstructure. A Microstructure object can be obtained directly or created using the R1, R2, off_resonance, diffusivity, and geometry flags.

timestep::Real: Timestep of the spin random walk through the tissue in milliseconds (default: 0.5).

Running the simulation

To run a Snapshot through the simulations you can use one of the following functions:

evolve: evolves the spins in the snapshot until a single given time and returns that state in a new Snapshot.

trajectory: returns full spin trajectory (recommended only for small number of spins)

signal: returns signal variation over time

readout: returns the snapshots at the sequence readouts.

MCMRSimulator.Snapshot — Type

Represents the positions and orientations of multiple Spin objects at a specific time.

Note that times are in milliseconds and positions in micrometer. The equilibrium longitudinal spin (after T1 relaxation) is always 1.

Useful constructors

Snapshot(positions; time=0., longitudinal=1., transverse=0., phase=0., nsequences=1)
Snapshot(nspins::Integer; time=0., longitudinal=1., transverse=0., phase=0., nsequences=1)

Creates a new Snapshot at the given time with perfectly longitudinal spins initialised for simulating nsequences sequences. This initial spin locations are given by positions (Nx3 matrix or sequence of vectors of size 3). Alternatively the number of spins can be given in which case the spins are randomly distributed in a 1x1x1 mm box centered on the origin.

Snapshot(snap::Snapshot{1}, nsequences)

Replicates the positions and orientations for a single sequence in the input snapshot across nsequences.

Extracting summary information

longitudinal(snapshot) to get the nsequences spin magnitudes in the z-direction (equilibrium of 1) averaged over all spins

transverse(snapshot) to get the nsequences spin magnitudes in the x-y-plane averaged over all spins

phase(snapshot) to get the nsequences spin angles in x-y plane (in degrees) averaged over all spins

orientation(snapshot) to get a (nsequencesx3) matrix with the spin orientations in 3D space averaged over all spins

SpinOrientation(snapshot) to get a nsequences vector of [SpinOrientation] objects with the average spin orientation across all spins

position.(snapshot) to get a the position for each spin in a vector (no averaging applied)

Information for a single sequence can be extracted by calling get_sequence first.

MCMRSimulator.Sphere — Type

Sphere([radius[, location]])

Creates a hollow sphere with a radius of radius micrometer (default 1 micrometer) at the given location (default: origin). Generate spheres using spheres.

MCMRSimulator.Spin — Type

Spin particle with a position and nsequences spin orientations (stored as SpinOrientation).

A random number generator is stored in the Spin object as well, which will be used for evolving the spin into the future in a reproducible manner.

Constructors

Spin(;nsequences=1, position=[0, 0, 0], longitudinal=1., transverse=0., phase=0.)

Creates a new spin with nsequences identical spin orientations (given by longitudinal, transverse, and phase flags). The spin will start at given position.

Spin(reference_spin::Spin{1}, nsequences)

Create a new spin with the same position as reference_spin with the orientation of reference_spin replicated nsequences times.

Extracting spin information

longitudinal to get the nsequences spin magnitudes in the z-direction (equilibrium of 1)

transverse to get the nsequences spin magnitudes in the x-y-plane

phase to get the nsequences spin angles in x-y plane (in degrees)

orientation to get a (nsequencesx3) matrix with the spin orientations in 3D space

position to get a length-3 vector with spin location

MCMRSimulator.SpinOrientation — Type

SpinOrientation(longitudinal, transverse, phase)

The spin orientation. Usually created as part of a Spin object.

SpinOrientation(snapshot::Snapshot)

Returns the average spin orientations of all Spin objects in the Snapshot.

This information can be extracted using:

longitudinal to get the spin in the z-direction (equilibrium of 1)

transverse to get the spin in the x-y-plane

phase to get the spin angle in x-y plane (in degrees)

orientation to get the spin orientation as a length-3 vector

MCMRSimulator.Spiral — Type

Spiral(inner, outer; theta0=0., thickness=0.014)

Creates a spiral starting from inner to outer radius. The distance between subsequent wraps is thickness micrometer. The angle of the start of the wraps is theta0.

MCMRSimulator.StepWiseGradients — Type

create_gradients([(<start_time>, <gradient>), ...], interpolate=:step)

Models MRI gradients as a sequence of stepwise functions.

MCMRSimulator.TransformObstruction — Type

TransformObstruction(base_type, args...; positions=nothing, repeats=0., rotation=I(3), lorentz_radius=10., kwargs...)
TransformObstruction(base; positions=nothing, repeats=0., rotation=I(3), lorentz_radius=10.)

Transforms the BaseObstruction objects in base in one of several ways:

Apply the shifts defined by positions (none applied by default).

Infinitely repeat the base obstruction every repeats.

Rotate the base obstructions by rotation.

The BaseObstruction objects an already be initialised. If a BaseObstruction type is given instead it will be initialised using args and kwargs not used by TransformObstruction.

MCMRSimulator.Wall — Type

Wall()

Wall stretching to infinite along two dimensions. Generate walls using walls.

Base.Libc.time — Method

time(snapshot)
time(sequence_component)
time(sequence, sequence_index)

Returns the time in milliseconds that a snapshot was taken or that a sequence component will have effect.

MCMRSimulator.annuli — Method

annuli(inner, outer; myelin=false, chi_I=-0.1, chi_A=-0.1, positions=[0, 0], repeats=[Inf, Inf], rotation=I(3)

Creates one or more Annulus with given inner and outer radii. Myelinated annuli can be created by setting the myelin to true. All parameters can be either a single value or a vector of values.

The positions, repeats, and rotation control the annulus position and orientation and is explained in more detail in Defining the geometry. Additional keyword arguments are available to set generic obstruction settings as described in ObstructionProperties.

MCMRSimulator.apply — Method

apply(sequence_component, spin_orientation[, position])

Applies given sequence component to the spin orientation. Returns a new SpinOrientation. Some pulses (e.g., InstantGradient) require positional information as well.

apply(sequence_components, spin)
apply(sequence_components, snapshot)

Apply all sequence components to the spin orientation in the Spin or to all the spins in Snapshot (the return type matches this input type). Sequence components (see Sequence) can be nothing if there is no sequence component at this time.

MCMRSimulator.color — Method

color(orient::SpinOrientation; saturation=1.)

Returns a color representing the spin orientation in the transverse (x-y) plane. Brighter colors have a larger transverse component, so that spins with no transverse component are black. The actual color encodes the spin orientation.

MCMRSimulator.corners — Method

corners(bb::BoundingBox)

Returns a vector of all corners of the BoundingBox.

MCMRSimulator.correct_collisions — Method

correct_collision(movement, geometry)

Splits the given movement from point A to point B into multiple steps that bounce off the given obstructions. This function assumes perfect reflection rather than the diffuse reflection used in draw_step. It is used to test the collision detection and resolution, but not actually used in the simulations.

MCMRSimulator.create_gradients — Method

create_gradients([(<start_time>, <gradient>), ...], TR; interpolate=:linear)

Models MRI gradients as a sequence of linear functions. interpolate can be set to :step to model the gradients as stepwise functions.

MCMRSimulator.cylinders — Method

cylinders(radii; g_ratio=1, chi_I=-0.1, chi_A=-0.1, positions=[0, 0], repeats=[Inf, Inf], rotation=I(3)

Creates one or more Cylinders with given radius (or vector of radii). Myelinated cylinders can be created by setting the g_ratio to a different value that 1. All parameters can be either a single value or a vector of values. The positions, repeats, and rotation control the cylinder position and orientation and is explained in more detail in Defining the geometry. Additional keyword arguments are available to set generic obstruction settings as described in ObstructionProperties.

MCMRSimulator.derive_qval_time — Method

derive_qval_time(max_diffusion_time; bval=nothing, diffusion_time=nothing, qval=nothing)

Derives b-value, diffusion time, and q-value from each other assuming short pulses

Returns a tuple with the q-value and diffusion time.

The strength of the diffusion weighting can be defined in 3 ways:

not setting anything or only setting the diffusion time: in this case no diffusion weighting is applied (b=0).

setting only the b-value or q-value: in this case a diffusion_time of TE/2 is assumed.

setting two out of three of the b-value, diffusion_time, and qval. The last one will be calculated.

If all three are defined an AssertionError is raised if they do not agree with each other.

Assumes equation of $b = q^2 \Delta$, where

bval is the diffusion-weighted strength (b-value) in $ms/um^2$.

qval is the gradient applied due to the diffusion-weighted gradients in the spin phase field ($rad/\mu m$). For a square pulse this is computed as $\gamma G \delta$.

diffusion_time ($\Delta$) is the diffusion time defined for infinitely short pulses as the time between the diffusion-weighted gradients (in $ms$).

MCMRSimulator.detect_collision — Method

detect_collision(movement, obstructions[, previous])

Returns a Collision object if the given movement crosses any obstructions. The first collision is always returned. If no collision is detected, empty_collision will be returned

MCMRSimulator.draw_step — Method

draw_step(current::Spin, diffusivity::Float, timestep::Float[, geometry::Obstructions])

Returns a new spin with the state of the particle after timestep:

draws the next location of the particle after timestep with given diffusivity. If provided, this displacement will take into account the obstructions in geometry.

the spin orientation might be affected by collisions with the obstructions in geometry.

MCMRSimulator.dwi — Method

dwi(; TE=80., TR=2000., bval/qval/gradient_strength, diffusion_time=TE/2, gradient_duration, readout_time=0., scanner=Scanner(B0=3.), orientation = [1., 0, 0])

Creates a diffusion-weighted pulsed gradient spin echo MRI sequence

The timings of the RF pulses is set by TE and TR. The gradient timings will also be affected by gradient_duration, diffusion_time, scanner (which sets the ramp time) and readout_time:

By default the gradient durations are set to the maximum value possible within the echo time (TE) keeping in mind the time needed for the MR readout (readout_time) and the time needed to ramp to the maximum gradient strength (set by the scanner).

When gradient_duration is set to 0, the gradient pulses are assumed to be instanteneous (i.e., using InstantGradient). The time between these instant gradients can be set using diffusion_time (defaults to TE/2).

The strength of the diffusion gradients is set by one of bval, qval, or gradient_strength. If this strength exceeds the maximum allowed for the scanner an AssertionError is raised. The gradient orientation is set by orientation.

MCMRSimulator.dyad_snapshot! — Method

dyad_snapshot(plot_plane, snapshot; dyadlength=0.1, arrowsize=0.1, color=:black, sequence=1)

Plots the spins in the Snapshot projected onto given PlotPlane as arrows.

MCMRSimulator.dyad_snapshot — Method

dyad_snapshot(plot_plane, snapshot; dyadlength=0.1, arrowsize=0.1, color=:black, sequence=1)

Plots the spins in the Snapshot projected onto given PlotPlane as arrows.

MCMRSimulator.evolve — Method

evolve(snapshot, simulation[, new_time])

Evolves the Snapshot through the Simulation to a new time. Returns a Snapshot at the new time, which can be used as a basis for further simulation. By default it will simulate till the start of the next TR.

MCMRSimulator.evolve_to_time — Method

evolve_to_time(snapshot, simulation, new_time)

Evolves the full Snapshot through the Simulation to the given new_time. Multi-threading is used to evolve multiple spins in parallel. This is used internally when calling any of the snapshot evolution methods (e.g., evolve).

MCMRSimulator.evolve_to_time — Method

evolve_to_time(spin, simulation, current_time, new_time)

Evolve a single spin to the next time of interest. This takes into account both random diffusion of the spin's position and relaxation of the MR spin orientation. It is used internally when evolving Simulation objects.

MCMRSimulator.expand — Function

expand(bb::BoundingBox, ratio=1.)

Expand or shrink the BoundingBox so that it grows/shrinks by ratio along each dimension.

MCMRSimulator.field — Method

Construct a Field object to describe the spatial distribution of some parameter (e.g., R1, R2, off-resonance, diffusivity)

Constructors

field()

Sets the parameter to zero everywhere.

field(value::real)

Sets the parameter to a constant value everywhere.

field(gradient::PosVector, value::real)

Parameter is described by a spatial gradient.

MCMRSimulator.fit_time — Method

fit_time(gradient_duration, diffusion_time, ramp_time, readout_time, TE)

Arranges the timing of gradients given the duration of gradient, diffusion time, ramp time, readout time and TE. By default it will try to arrange to gradients symmetrically arround the 180 degree pulse. If not feasible, it will make readout right after the rephasing gradient and calculate the dephasing gradient's timing based on the rephasing gradient.

This function is not supposed to be called individually! So no check for the arguments will happen, assuming all have been checked by dwigradients1D and all will not be nothing.

MCMRSimulator.flip_angle — Method

flip_angle(pulse)

Returns the flip angle of the RF pulse in degrees

MCMRSimulator.get_gradient — Method

get_gradient([position, ]sequence/grad, time)
get_gradient([position, ]sequence/grad, t1, t2)

Get gradient strength generated by grad (MRGradients object) at a specific time or averaged over a period between t1 and t2. If no position is provided, the gradient is returned as a length-3 vector in units of mT/m. If a position is provided, the gradient is returned at that position as a float quantifying the off-resonance field in units of microTesla.

MCMRSimulator.get_sequence — Method

get_sequence(spin, sequence_index)
get_sequence(snapshot, sequence_index)

Extracts the spin orientation corresponding to a specific sequence, where the sequence index uses the order in which the sequences where provided in the Simulation.

MCMRSimulator.get_times — Method

get_times(simulation::Simulation, t_start, t_end)
get_times(time_controller::TimeController, t_start, t_end, sequences, diffusivity)

Computes the timepoints at which the simulation will be evaluated when running from tstart to tend.

The following timepoints will always be included

Any multiple of the TR

Any RF pulse

Any change in the gradient strength

Additional timepoints will be added to ensure that:

at any time the timestep is larger than time_controller.gradient_precision times $D^{-1/3} (G \gamma)^{-2/3}$.

between t_start and t_end there are at least time_controller.sample_displacement timepoints.

between any control point of the sequence gradient there are at least time_controller.sample_displacement timepoints.

between an RF pulse and a subsequent RF pulse or tend there are at least `timecontroller.sampleoffresonance` timepoints.

MCMRSimulator.image_snapshot! — Method

image_snapshot(plot_plane, snapshot; vectorlength=0.1, arrowsize=0.1, color=:black, sequence=1)
image_snapshot!(plot_plane, snapshot; vectorlength=0.1, arrowsize=0.1, color=:black, sequence=1)

Plots the spins in the Snapshot projected onto given PlotPlane. The average spin orientation across the plot plane is plotted using the colour coding from color.

MCMRSimulator.image_snapshot — Method

image_snapshot(plot_plane, snapshot; vectorlength=0.1, arrowsize=0.1, color=:black, sequence=1)
image_snapshot!(plot_plane, snapshot; vectorlength=0.1, arrowsize=0.1, color=:black, sequence=1)

Plots the spins in the Snapshot projected onto given PlotPlane. The average spin orientation across the plot plane is plotted using the colour coding from color.

MCMRSimulator.isinside — Method

isinside(obstructions/bounding_box, position)
isinside(obstructions/bounding_box, spin)
isinside(obstructions/bounding_box, snapshot)

Test whether the particles are inside any of the Obstruction objects (or in the BoundingBox).

MCMRSimulator.load_mesh — Method

load_mesh(file)

Loads a Mesh from a file.

Currently only PLY files are supported (see ply_from_mesh)

MCMRSimulator.longitudinal — Function

transverse(spin)
transverse(snapshot)

Returns the longitudinal magnitude of the spin (i.e., magnitude aligned with the magnetic field) for a single particle (Spin) or averaged across a group of particles in a [Snapshot]. When orientations for multiple sequences are available an array of longitudinal values is returned with a value for each sequence.

MCMRSimulator.lorentz_off_resonance — Method

lorentz_off_resonance(obstructions, position, b0_field, repeat_dist, radius, nrepeats)

Computes the off-resonance field contribution of repeating compartments within a spherical or cylindrical Lorentz cavity.

MCMRSimulator.mesh_grid_intersection — Method

mesh_grid_intersection(shape, vertices, triangles)

Determines for each voxel in the grid with which triangles it intersects.

Algorithm

We follow the algorithm proposed by https://fileadmin.cs.lth.se/cs/Personal/TomasAkenine-Moller/code/triboxtam.pdf, which is based on the separating axis theorem.

We project each voxel in the grid and each triangle on a series of 9 axes:

the normals of the cube ([1, 0, 0], [0, 1, 0], [0, 0, 1])

the normal of the triangle

the cross products of any edge of the cube (same as normals) and the edges of the triangle (total of 9 tests)

If the cube and triangle do not overlap along any of these axes, they are non-overlapping. If they overlap along all axes, the triangle and cube overlap

MCMRSimulator.next_pulse — Method

next_pulse(sequence, current_time)

Find the index of the next pulse that will be applied. For an empty sequence 0 will be returned.

MCMRSimulator.normal — Method

normal(p1, p2, p3)

Computes the normal of a triangle formed by the three points

MCMRSimulator.off_resonance — Function

off_resonance(obstructions, position[, b0_field])

Computes the off-resonance field at the position due to the obstructions with the magnetic field orientation from b0_field.

MCMRSimulator.off_resonance — Method

off_resonance(cylinder, position, b0_field)

Computed by the hollow cylinder fiber model from Samuel Wharton, Richard Bowtell (2012).

MCMRSimulator.orientation — Function

orientation(spin)
orientation(snapshot)

Returns the spin orientation as a length-3 vector for a single particle (Spin) or averaged across a group of particles in a [Snapshot]. When orientations for multiple sequences are available an array of vectors is returned with a value for each sequence.

MCMRSimulator.perfect_dwi — Method

perfect_dwi(; TE=80., TR=2000., bval=nothing, diffusion_time=nothing, qval=nothing, orientation = [0, 0, 1.])

Creates a standard diffusion-weighted MRI sequence

This produces a pulsed gradient spin echo (PGSE) sequence with perfect RF pulses and infinitely short diffusion-weighted gradients.

The procedure for defining the diffusion-weighted gradient strength is defined by derive_qval_time. The orientation parameter only sets the gradient orientation, not the strength.

MCMRSimulator.phase — Function

phase(spin)
phase(snapshot)

Returns the phase in the x-y plane of the spin for a single particle (Spin) or averaged across a group of particles in a [Snapshot]. When orientations for multiple sequences are available an array of phase values is returned with a value for each sequence.

MCMRSimulator.phase — Method

phase(pulse)

Returns the angle in the x-y plane of the axis around with the RF pulse rotates in degrees

MCMRSimulator.plot_geometry! — Method

plot(plot_plane, geometry)
plot!(plot_plane, geometry)
plot_geometry(plot_plane, geometry)
plot_geometry!(plot_plane, geometry)

Plots the intersections of geometry in the PlotPlane.

MCMRSimulator.plot_geometry — Method

plot(plot_plane, geometry)
plot!(plot_plane, geometry)
plot_geometry(plot_plane, geometry)
plot_geometry!(plot_plane, geometry)

Plots the intersections of geometry in the PlotPlane.

MCMRSimulator.plot_off_resonance! — Method

plot_off_resonance(plot_plane, geometry)
plot_off_resonance(plot_plane, geometry)

Plots the off-resonance of geometry in the PlotPlane.

MCMRSimulator.plot_off_resonance — Method

plot_off_resonance(plot_plane, geometry)
plot_off_resonance(plot_plane, geometry)

Plots the off-resonance of geometry in the PlotPlane.

MCMRSimulator.plot_snapshot — Function

plot(plot_plane, snapshot; kwargs...)
plot!(plot_plane, snapshot; kwargs...)
plot_snapshot(plot_plane, snapshot; kwargs...)
plot_snapshot!(plot_plane, snapshot; kwargs...)

Plots the spins in the Snapshot projected onto given PlotPlane. Each spin is represented by an arrow showing the transverse component of the spin (for a total of the first ndyads spins). The average spin orientation across the plot plane is plotted using the colour coding from color.

This combines the plotting from image_snapshot and dyad_snapshot.

MCMRSimulator.plot_trajectory2d! — Method

plot(snapshots)
plot!(snapshots)
plot_trajectory3d(snapshots)
plot_trajectory3d!(snapshots)

Plots trajectory of the spins in a sequence of Snapshot objects.

MCMRSimulator.plot_trajectory2d — Method

plot(snapshots)
plot!(snapshots)
plot_trajectory3d(snapshots)
plot_trajectory3d!(snapshots)

Plots trajectory of the spins in a sequence of Snapshot objects.

MCMRSimulator.plot_trajectory3d! — Method

plot(snapshots)
plot!(snapshots)
plot_trajectory3d(snapshots)
plot_trajectory3d!(snapshots)

Plots trajectory of the spins in a sequence of Snapshot objects.

MCMRSimulator.plot_trajectory3d — Method

plot(snapshots)
plot!(snapshots)
plot_trajectory3d(snapshots)
plot_trajectory3d!(snapshots)

Plots trajectory of the spins in a sequence of Snapshot objects.

MCMRSimulator.ply_from_mesh — Method

ply_from_mesh(file)

Loads a Mesh from a PLY file. PLY stands for Polygon File Format (http://paulbourke.net/dataformats/ply/). PLY IO is handled by PlyIO.jl (https://github.com/JuliaGeometry/PlyIO.jl).

MCMRSimulator.position — Method

position.(s::Snapshot)

Returns all the positions of the spin particles as a vector of length-3 vectors.

MCMRSimulator.position — Method

position(s::Spin)

Returns the position of the spin particle as a vector of length 3.

MCMRSimulator.produces_off_resonance — Method

produces_off_resonance(simulation)
produces_off_resonance(geometry)

Tests whether any aspects of the simulation geometry will produce an off-resonance field.

MCMRSimulator.project — Function

project(position, transform::TransformObstruction)

Computes the position in the space of the obstructions wrapped by the TransformObstruction.

project(position, grid::GridShape)

Computes the voxel index for the position on the GridShape. This will return a result even if the point is outside of the grid. Use isinside(position, grid) to check that.

MCMRSimulator.project — Method

project(plot_plane, position)
project(plot_plane, snapshot)

Transforms the position (length-3 vector) or Snapshot to a space, where the PlotPlane lies in the x-y-plane centered on origin.

MCMRSimulator.project_geometry — Method

project_geometry(plot_plane, transform)

Projects the plane on the intrinsic plane of the obstructions deformed by transform.

MCMRSimulator.project_on_grid — Method

project_on_grid(plot_plane, snap, ngrid)

Spins from the Snapshot are projected onto the grid defined by PlotPlane in two ways:

along the normal spins are projected onto the plane from infinitely far (TODO: give finite extent)

in the other directions any spins are projected onto the plane using mod(position[1], sizex) and mod(position[2], sizey). This assumes that the geometry and field repeats itself ad infinitum beyond the PlotPlane (TODO: allow this assumption to be turned off).

In effect, this means that all spins are projected onto the PlotPlane. The average spin orientation in each grid cell is returned.

MCMRSimulator.random_annuli — Method

random_annuli(target_density; repeats, g_ratio=0.8, distribution=Distributions.Gamma, mean_radius=1., variance_radius=0.5, max_iter=1000, myelin=false, chi_I=-0.1, chi_A=-0.1, rotation=I(3))

Generate infinitely repeating box with non-overlapping annuli.

A rectangle with the size of repeats will be filled with annuli for a total surface density of target_density. The annulus outer radii will be drawn from the selected distribution (if not set, a Gamma distribution is used with given mean_radius and var_radius). An error is raised if no solution for non-overlapping annuli is found. The inner radius with respect to the outer radius is set by the g-ratio. Other annuli parameters (besides inner, outer, positions, and repeats) are identical as in mr.annuli.

MCMRSimulator.random_cylinders — Method

random_cylinders(target_density; repeats, distribution=Distributions.Gamma, mean_radius=1., variance_radius=0.5, max_iter=1000, g_ratio=1., chi_I=-0.1, chi_A=-0.1, rotation=I(3))

Generate infinitely repeating box with non-overlapping cylinders.

A rectangle with the size of repeats will be filled with cylinders for a total surface density of target_density. The cylinder radii will be drawn from the selected distribution (if not set, a Gamma distribution is used with given mean_radius and var_radius). An error is raised if no solution for non-overlapping cylinders is found. Other cylinder parameters (besides radii, shifts, and repeats) are identical as in mr.cylinders.

MCMRSimulator.random_gauss — Method

random_gauss(diffusivity, timestep)

Draws a random step from a Gaussian distribution with a variance of 2 * diffusivity * timestep.

MCMRSimulator.random_on_sphere — Method

random_on_sphere()

Draws a random orientation on the unit sphere as a length-3 vector

The z-orientation is drawn a random number between -1 and 1. The angle in the x-y plane is drawn as a random number between 0 and 2π. This results in an unbiased random distribution across the sphere.

MCMRSimulator.random_positions_radii — Method

random_positions_radii(box_size, target_density, n_dimensions; distribution=Gamma, mean=1., variance=1., max_iter=1000)

Randomly distributes circles or spheres in space.

Arguments:

box_size: Size of the infinitely repeating box of random positions

target_density: Final density of the circles/spheres. This density will only be approximately reached

n_dimensions: dimensionality of the space (2 for cicles; 3 for spheres)

distribution: distribution from which the radii are drawn (from Distributions.jl)

mean: mean of the gamma distribution (ignored if distribution explicitly set)

variance: variance of the gamma distribution (ignored if distribution explicitly set)

max_iter: maximum number of iterations to try to prevent the circles/spheres from overlapping. An error is raised if they still overlap after this number of iterations.

MCMRSimulator.random_spheres — Method

random_spheres(target_density; repeats, distribution=Distributions.Gamma, mean_radius=1., variance_radius=0.5, max_iter=1000, rotation=I(3))

Generate infinitely repeating box with non-overlapping spheres.

A box with the size of repeats will be filled with spheres for a total volume density of target_density. The sphere radii will be drawn from the selected distribution (if not set, a Gamma distribution is used with given mean_radius and var_radius). An error is raised if no solution for non-overlapping spheres is found. Other sphere parameters (besides radii, positions, and repeats) are identical as in mr.spheres.

MCMRSimulator.random_spirals — Method

random_spirals(target_density; repeats, g_ratio=0.8, distribution=Distributions.Gamma, mean_radius=1., variance_radius=0.5, max_iter=1000, rotation=I(3))

Generate infinitely repeating box with non-overlapping spirals.

A rectangle with the size of repeats will be filled with spirals for a total surface density of target_density. The spiral outer radii will be drawn from the selected distribution (if not set, a Gamma distribution is used with given mean_radius and var_radius). An error is raised if no solution for non-overlapping annuli is found. The inner radius with respect to the outer radius is set by the g-ratio. Other spiral parameters (besides inner, outer, positions, and repeats) are identical as in mr.spirals.

MCMRSimulator.ray_grid_intersections — Method

ray_grid_intersections([grid, ]origin, destination)

Computes all voxels crossed by a ray between origin and destination with a GridShape (default infinitely extending 1x1x1 grid). Both origin and destination are length-3 vectors. The returned object is an iterator returning a tuple with:

3-length vector with the voxel that we are crossing through

Float with the time the ray entered voxel (0=origin, 1=destination)

3-length vector with position within voxel that the ray entered (i.e., numbers between 0 and 1)

Float with the time the ray left the voxel (0=origin, 1=destination)

3-length vector with position within voxel that the ray left (i.e., numbers between 0 and 1)

MCMRSimulator.readout — Method

readout(snapshot, simulation)

Evolves the spins in the Snapshot through the Simulation. Returns the Snapshot at every Readout in the simulated sequences during a single TR. If no TR is explicitly selected, it will return the current TR if the snapshot has not passed any readouts and the next TR otherwise.

MCMRSimulator.relax — Function

relax(spin_orientation, environment, timestep::Real, gradient_off_resonance=0., B0=3.)

Relaxes the SpinOrientation within the R1, R2, and off-resonance given by the LocalEnvironment over given timestep

MCMRSimulator.rotate_bvec — Method

rotate_bvec(gradients, bvec; orig_bvec=nothing)

Rotates the gradients in gradients to align with bvec. If not provided, the original gradient orientation is estimated as the first eigenvector of the b-tensor.

MCMRSimulator.scatter_snapshot! — Method

plot(snapshot)
plot!(snapshot)
scatter_snapshot(snapshot)
scatter_snapshot!(snapshot)

Plots the spin positions in the Snapshot in 3D color coded by the spin's orientation (see color).

MCMRSimulator.scatter_snapshot — Method

plot(snapshot)
plot!(snapshot)
scatter_snapshot(snapshot)
scatter_snapshot!(snapshot)

Plots the spin positions in the Snapshot in 3D color coded by the spin's orientation (see color).

MCMRSimulator.sequence_plot — Function

plot(sequence)
plot!(sequence)
plot_sequence(sequence)
plot_sequence!(sequence)

Creates a visual representation of a Sequence diagram.

MCMRSimulator.signal — Method

signal(snapshot, simulation, times=[TR])

Evolves the Snapshot through the Simulation and outputs the total signal at the requested times. To get the full snapshot at each timepoint use trajectory. Returns a vector of SpinOrientation object with the total signal at each time in times.

MCMRSimulator.spheres — Method

spheres(radii; positions=[[0, 0, 0]], repeats=[Inf, Inf, Inf], rotation=I(3))

Creates one or more Spheres with given radius (or vector of radii). The positions, repeats, and rotation control the sphere positions and is explained in more detail in Defining the geometry. Additional keyword arguments are available to set generic obstruction settings as described in ObstructionProperties.

MCMRSimulator.spiral_theta — Method

spiral_theta(spiral, position; assume_inner)

Computes the location of the particle within the spiral. This function returns 0 if the particle has just entered the inner part of the spiral. Each wrap further outwards will add 2π to the result.

When the particle is at the spiral edge, assume_inner should be set to true if the particle should be considered just on the inner side of the surface or to false if it should be considered just on the outer surface.

MCMRSimulator.spirals — Method

spirals(inner, outer; theta0=0., thickness=0.014, myelin=false, chi_I=-0.1, chi_A=-0.1, positions=[0, 0], repeats=[Inf, Inf], rotation=I(3)

Creates one or more Spiral. Spirals range from inner to outer radii starting at an angle of theta0. Each wrap has a thickness of thickness micrometers. Inner/outer cylinders can be added using respectively the inner_cylinder and outer_cylinder flags (default: only inner cylinder).

Myelinated spirals can be created by setting the myelin to true. All parameters can be either a single value or a vector of values.

The positions, repeats, and rotation control the annulus position and orientation and is explained in more detail in Defining the geometry. Additional keyword arguments are available to set generic obstruction settings as described in ObstructionProperties.

MCMRSimulator.total_susceptibility — Method

total_susceptibility(obstruction)

Computes the total magnetic susceptibility of an obstructions. This might be integrated over the surface or the volume depending on the dimensionality of the obstruction.

MCMRSimulator.trajectory — Method

trajectory(snapshot, simulation, times=[TR])

Evolves the Snapshot through the Simulation and outputs at the requested times. Returns a vector of Snapshot objects with the current state of each time in times. When you are only interested in the signal at each timepoint, use signal instead.

MCMRSimulator.transfer — Method

transfer(spin1, spin2, fraction)
transfer(orientation, obstruction)

Transfers fraction of the total spin orientation between spin1 and spin2. Alternatively transfers spin between spin orientation and obstruction.

MCMRSimulator.transverse — Function

transverse(spin)
transverse(snapshot)

Returns the transverse spin (i.e., magnitude in the plane perpendicular to the magnetic field) for a single particle (Spin) or averaged across a group of particles in a [Snapshot]. When orientations for multiple sequences are available an array of transverse values is returned with a value for each sequence.

MCMRSimulator.walls — Method

walls(positions=0, repeats=Inf, rotation=I(3))

Creates one or more Walls. The positions, repeats, and rotation control the wall position and orientation and is explained in more detail in Defining the geometry. Additional keyword arguments are available to set generic obstruction settings as described in ObstructionProperties.

MakieCore.plot! — Method

dyad_snapshot(plot_plane, snapshot; dyadlength=0.1, arrowsize=0.1, color=:black, sequence=1)
dyad_snapshot(plot_plane, snapshot; dyadlength=0.1, arrowsize=0.1, color=:black, sequence=1)

Plots the spins in the Snapshot projected onto given PlotPlane. Each spin is represented by an arrow showing the transverse component of the spin.